Split up HCA reducers (#14984)
So that the Dashboard can pull in only the part that it needs.
