+++ src/applications/personalization/dashboard/reducers/index.js
@@ -1,10 +1,10 @@
 import claimsAppeals from '~/applications/claims-status/reducers';
+import hcaEnrollmentStatus from '~/applications/hca/reducers/hca-enrollment-status-reducer';
 import prescriptions from './prescriptions';
 import recipients from './recipients';
 import folders from './folders';
 import preferences from '../../preferences/reducers';
 import appointments from '../../appointments/reducers';
-import { hcaEnrollmentStatus } from 'applications/hca/reducer';
 
 import { combineReducers } from 'redux';
