+++ src/applications/personalization/profile/reducers/index.js
@@ -1,5 +1,5 @@
 import vapService from '@@vap-svc/reducers';
-import { hcaEnrollmentStatus } from 'applications/hca/reducer';
+import hcaEnrollmentStatus from '~/applications/hca/reducers/hca-enrollment-status-reducer';
 import vaProfile from './vaProfile';
 
 export default {
